+++ CHANGELOG.rst
@@ -3,11 +3,14 @@
 
 Added:
 ^^^^^^
-- Added option to disable hard filter of variants in matched normal
+* Added option to disable hard filter of variants in matched normal https://github.com/Clinical-Genomics/BALSAMIC/pull/1509
 
 Changed:
 ^^^^^^^^
-- Reworked bcftools filters
+* Reworked bcftools filters https://github.com/Clinical-Genomics/BALSAMIC/pull/1509
+* Renamed high_normal_tumor_af_frac to in_normal https://github.com/Clinical-Genomics/BALSAMIC/pull/1509
+* check to verify sample sex for all workflows https://github.com/Clinical-Genomics/BALSAMIC/pull/1516
+
 
 Removed:
 ^^^^^^^^